In ConfigList FILE, LIST, MEDIA, PVR and VIDEO to Select
To these buttons the menu button was assigned and it seems other images
put the selection screen on the menu button. We made the menu button
flexible for other functionality. But togther with the menu button also
the FILE, LIST, MEDIA, PVR or VIDEO button were also assigned to this
menu button. By assigning them to 'select' you have actually the same
functionality as the OK button to these buttons so in plugins that are
not adjusted yet to the new configlist stuff you're still able to get to
the selection list via these buttons, while the menu button stays
configurable.