Finetune scrollbar where sometimes text does not fully fit
The height of the text was determine previously when there was no scroll
bar. The scrollbar makes the text box smaller. So sometimes we missed
'height' so not all text was shown. Now we always reserve place for the
vertical scrollbar. So the height is always calculated correctly.

In addition I introduced a scrollbar margin attribute (default scrollbar
width 20, scrollbar margin 10)

Also finetune a onelayer as with an if the test for the split is more
clear I suppose (long long time ago I used these kind of 'and' things,
but not anymore).