Resolve issue with changes in Selection window in ScanSetup
When you change a setting via the OK button where you get the selection
box then after you selected something the setup list was not build so
the change did not had any effect. This happened in the Satfinder plugin
but also in on the ScanSetup screen.

This needs to be cherry picked also in Release 9.1