Get rid of the confusement between model and machine in boxinfo
I added now two variables machine which is actually the old name where
the image is build from and model. Some BoxInfo/SystemInfo stuff needs
more details and where needed it can later be derived from the variable
model. What was the variable model was previously machine from
enigma.info.

The current model variable is not used yet as only for some values it
might needed. So as soon the model in enigma.info get adapted this can
be adjusted