Only change references when parental check is passed
After the last changes in Navigation.py references were changed earlier
even before the parental check was passed. So e.g. when you select a
parental control channel and you need to enter a pincode the reference
was already set to the new channel so after the pin code entering
enigma2 did not zap to the channel. By moving the change of references
after the parental check this was resolved.